Improve pip auto installer
Use requirements.txt, add auto updating.
