+++ requirements.txt
@@ -1,5 +1,6 @@
+bitarray~=2.7.6
 imageio~=2.26.0
 numpy~=1.22
 PyQt5~=5.15.4
-future~=0.18.2
+PyQt5-Frameless-Window~=0.2.8
 vdf~=3.4
